Struct spades::card::set::Set

pub struct Set { /* fields omitted */ }

A set type for cards.

Methods

impl Set

pub fn full() -> Self

Creates a set containing every card.

pub fn suite(suite: Suite) -> Self

Creates a set containing every card of a suite.

pub fn insert(&mut self, card: Card) -> bool

Inserts a card into a set.

Returns true if the insert took place. Returns false if nothing happened due to the card already being inserted.

pub fn remove(&mut self, card: Card) -> bool

Removes a card from a set.

Returns true if the removal took place. Returns false if nothing happened due to the card not being in the set.

pub fn clear(&mut self)

Removes every card from this set.

pub fn contains(self, card: Card) -> bool

Gets if a card is in this set.

pub fn is_empty(self) -> bool

Gets if this set is empty.

pub fn len(self) -> usize

Gets the number of cards in this set.

pub fn iter(self) -> Iterator

Creates an iterator over all of the cards in this set.

The iteration order is from lowest to highest index. This means that all cards in the same suite will be ordered from lowest value to highest value.