Module spades::scoring

Contains functions and types that store and give information about team's scores.

Structs

Score
Represents a team's score.

TeamRoundResult
Contains a team's bid and number of tricks taken in a round.

Enums

Bid
A player's bid.

Functions

get_bid_value
Gets the value of a team's bid.

get_winning_team_index
Gets the index of the winning team.